requirements.txt
aiodns==3.0.0
aiofiles==0.8.0
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
APScheduler==3.6.3
astpretty==2.1.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
bandit==1.7.4
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
emoji==1.7.0
eradicate==2.1.0
flake8==4.0.1
flake8-annotations-complexity==0.0.7
flake8-awesome==1.3.0
flake8-bandit==3.0.0
flake8-breakpoint==1.1.0
flake8-bugbear==22.4.25
flake8-builtins==1.5.3
flake8-comprehensions==3.10.0
flake8-eradicate==1.2.1
flake8-expression-complexity==0.0.11
flake8-if-expr==1.0.4
flake8-isort==4.1.1
flake8-logging-format==0.6.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-print==5.0.0
flake8-pytest==1.3
flake8-pytest-style==1.6.0
flake8-requirements==1.5.5
flake8-return==1.1.3
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.8.0
iniconfig==1.1.1
isort==5.10.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
packaging==21.3
pbr==5.9.0
pep8-naming==0.13.0
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.19.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlitedict==2.0.0
stevedore==3.5.0
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
